pyproject.toml
[tool.poetry]
name = "murmur-decoder-bot"
version = "0.1.0"
description = "Telegram bot that transcribes voice messages in many languages using OpenAI Whisper API"
authors = ["Dima Krutolianov <DimaKrutDK@ya.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {extras = ["webhooks"], version = "^20.6"}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
openai = "^1.2.2"
sentry-sdk = "^1.34.0"
google-cloud-logging = "^3.11.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
